more udpates

+++ xarinit.php
@@ -41,7 +41,7 @@ function publications_init()
             body2 TEXT,
             body3 TEXT,
             notes TEXT,
-            parent_id tinyint NOT NULL DEFAULT '0',
+            parent_id integer unsigned NOT NULL DEFAULT '0',
             pubtype_id tinyint NOT NULL DEFAULT '1',
             pages INT UNSIGNED NOT NULL DEFAULT '1',
             locale varchar(20) NOT NULL DEFAULT '',
